client/subdomain_scan/domaininfo/requirements.txt
celery
redis
dnspython